requirements.txt
langchain
langchain-community
langchain-google-genai
python-dotenv
pydantic>=2
pandas
matplotlib
sqlalchemy
